labx/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labx"
version = "0.1.0"
description = "Lab Report Extraction & Analysis Engine — hospital-grade, async, vision-powered"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "MedWard Team" }]

dependencies = [
    "anthropic>=0.39.0,<1.0",
    "pydantic>=2.6,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "typer[all]>=0.12,<1.0",
    "httpx>=0.27,<1.0",
    "python-multipart>=0.0.9",
    "structlog>=24.1,<25.0",
    "tenacity>=8.3,<10.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "mypy>=1.10",
]
observability = [
    "opentelemetry-api>=1.24",
    "opentelemetry-sdk>=1.24",
    "prometheus-client>=0.20",
]

[project.scripts]
labx = "labx.cli.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["labx"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
